more fixes. trimming to one luis call per turn.
