+++ samples/javascript_nodejs/51.cafe-bot/bot.js
@@ -3,13 +3,9 @@
 
 const { ActivityTypes, CardFactory, MessageFactory } = require('botbuilder');
 const { DialogSet } = require('botbuilder-dialogs');
-const { LuisRecognizer } = require('botbuilder-ai');
 const { OnTurnProperty } = require('./dialogs/shared/stateProperties');
 const { WelcomeCard } = require('./dialogs/welcome');
 const { MainDispatcher } = require('./dialogs/dispatcher');
-
-// LUIS service type entry in the .bot file for dispatch.
-const LUIS_CONFIGURATION = 'cafeDispatchModel';
 
 // State properties
 const ON_TURN_PROPERTY = 'onTurnStateProperty';
@@ -42,16 +38,6 @@ class CafeBot {
         this.onTurnAccessor = conversationState.createProperty(ON_TURN_PROPERTY);
         this.dialogAccessor = conversationState.createProperty(DIALOG_STATE_PROPERTY);
 
-        // add recognizer
-        const luisConfig = botConfig.findServiceByNameOrId(LUIS_CONFIGURATION);
-        if (!luisConfig || !luisConfig.appId) throw new Error(`Cafe Dispatch LUIS model not found in .bot file. Please ensure you have all required LUIS models created and available in the .bot file. See readme.md for additional information.\n`);
-        this.luisRecognizer = new LuisRecognizer({
-            applicationId: luisConfig.appId,
-            azureRegion: luisConfig.region,
-            // CAUTION: Its better to assign and use a subscription key instead of authoring key here.
-            endpointKey: luisConfig.subscriptionKey
-        });
-
         // add main dispatcher
         this.dialogs = new DialogSet(this.dialogAccessor);
         this.dialogs.add(new MainDispatcher(botConfig, this.onTurnAccessor, conversationState, userState));
@@ -73,7 +59,7 @@ class CafeBot {
         case ActivityTypes.Message:
             // Process on turn input (card or NLP) and gather new properties
             // OnTurnProperty object has processed information from the input message activity.
-            let onTurnProperties = await this.detectIntentAndEntities(turnContext);
+            let onTurnProperties = await this.detectIntentAndEntitiesFromCardInput(turnContext);
             if (onTurnProperties === undefined) break;
 
             // Set the state with gathered properties (intent/ entities) through the onTurnAccessor
@@ -104,18 +90,16 @@ class CafeBot {
         await this.userState.saveChanges(turnContext);
     }
     /**
-     * Async helper method to get on turn properties from cards or NLU using https://LUIS.ai
+     * Async helper method to get on turn properties from cards 
      *
      * - All cards for this bot -
      *   1. Are adaptive cards. See https://adaptivecards.io to learn more.
      *   2. All cards include an 'intent' field under 'data' section and can include entities recognized.
-     * - Bot also uses a dispatch LUIS model that includes trigger intents for all dialogs.
-     *   See ./dialogs/dispatcher/resources/cafeDispatchModel.lu for a description of the dispatch model.
      *
      * @param {TurnContext} turn context object
      *
      */
-    async detectIntentAndEntities(turnContext) {
+    async detectIntentAndEntitiesFromCardInput(turnContext) {
         // Handle card input (if any), update state and return.
         if (turnContext.activity.value !== undefined) {
             return OnTurnProperty.fromCardInput(turnContext.activity.value);
@@ -129,15 +113,9 @@ class CafeBot {
 
         // Nothing to do for this turn if there is no text specified.
         if (turnContext.activity.text === undefined || turnContext.activity.text.trim() === '') {
-            return;
+            return undefined;
         }
-
-        // Call to LUIS recognizer to get intent + entities
-        const LUISResults = await this.luisRecognizer.recognize(turnContext);
-
-        // Return new instance of on turn property from LUIS results.
-        // Leverages static fromLUISResults method
-        return OnTurnProperty.fromLUISResults(LUISResults);
+        return new OnTurnProperty();
     }
     /**
      * Async helper method to welcome all users that have joined the conversation.
