+++ samples/javascript_nodejs/51.cafe-bot/cognitiveModels/cafeDispatchModel.luis
@@ -247,6 +247,11 @@
       "entities": []
     },
     {
+      "text": "do you offer free wifi?",
+      "intent": "QnA",
+      "entities": []
+    },
+    {
       "text": "Free parking?",
       "intent": "QnA",
       "entities": []
