+++ samples/javascript_nodejs/51.cafe-bot/dialogs/bookTable/resources/turn-N.lu
@@ -51,6 +51,13 @@
 
 # Cancel
 - [Cancel intent](../../dispatcher/resources/cancel.lu#Cancel)
+- cancel the reservation
+- stop the booking
+- I do not wish to proceed with the reservation
+- I do not want the table anymore
+- I dont need the table anymore
+- sorry, i dont need the reservation
+- sorry, i dont need the booking anymore
 
 > Add reference to top level dispatch intents to none intent so this child model does not fire
 # Interruptions
